% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Square_fun.R
\name{square_fun}
\alias{square_fun}
\title{Trial Square Function}
\usage{
square_fun(x)
}
\arguments{
\item{x}{A numeric value}
}
\value{
A numeric value of the squared of number
}
\description{
Trial Square Function
}
\examples{
x <- 4
square_fun(4)
}
